==> code/00_ingest_data.R <==

pacman::p_load(
  tidyverse,
  fst
)

df_raw <- read_csv("https://raw.githubusercontent.com/NeilPaine538/MLB-WAR-data-historical/master/jeffbagwell_war_historical.csv") %>% 
  glimpse()

write_fst(df_raw, "data/raw/all_data.fst")

df_payroll <- data.frame(
  stringsAsFactors = FALSE,
       check.names = FALSE,
              Club = c("Dodgers","Yankees","Mets",
                       "Phillies","Astros","R. Sox","Padres","Angels",
                       "Cardinals","Nationals","Giants","B. Jays","Cubs","Braves",
                       "W. Sox","Reds","Twins","Rockies","Brewers",
                       "Rangers","D-Backs","Royals","Mariners","A's","Tigers",
                       "Rays","Marlins","Orioles","Guardians","Pirates"),
            `2021` = c("$262,108,948","$203,628,978",
                       "$198,818,899","$192,366,818","$189,586,134",
                       "$187,356,913","$184,125,634","$182,925,936","$168,923,070",
                       "$168,252,361","$165,869,140","$152,261,601",
                       "$152,189,618","$148,497,447","$141,908,402","$131,309,602",
                       "$125,983,176","$119,908,394","$105,991,594",
                       "$104,310,488","$99,951,436","$94,475,853","$93,650,276",
                       "$90,402,244","$86,047,409","$76,856,333","$60,877,424",
                       "$59,054,021","$53,010,498","$50,337,389"),
            `2019` = c("$205,147,830","$225,978,235",
                       "$146,217,185","$160,974,755","$169,831,987",
                       "$228,403,210","$104,000,709","$165,595,376","$178,872,485",
                       "$205,283,436","$182,810,115","$120,288,261",
                       "$220,307,173","$136,147,636","$97,304,607","$135,364,998",
                       "$125,205,980","$151,889,647","$132,645,682",
                       "$135,307,635","$109,903,547","$101,984,743","$139,064,933",
                       "$95,733,594","$117,097,633","$67,443,534","$75,671,476",
                       "$82,784,110","$123,127,347","$78,286,587"),
            Change = c("$56,961,118","($22,349,257)",
                       "$52,601,714","$31,392,063","$19,754,147",
                       "($41,046,297)","$80,124,925","$17,330,560","($9,949,415)",
                       "($37,031,075)","($16,940,975)","$31,973,340",
                       "($68,117,555)","$12,349,811","$44,603,795","($4,055,396)",
                       "$777,196","($31,981,253)","($26,654,088)","($30,997,147)",
                       "($9,952,111)","($7,508,890)","($45,414,657)",
                       "($5,331,350)","($31,050,224)","$9,412,799","($14,794,052)",
                       "($23,730,089)","($70,116,849)","($27,949,198)"),
         `%.(+/-)` = c("28%","-10%","36%","20%",
                       "12%","-18%","77%","10%","-6%","-18%","-9%","27%",
                       "-31%","9%","46%","-3%","1%","-21%","-20%","-23%",
                       "-9%","-7%","-33%","-6%","-27%","14%","-20%",
                       "-29%","-57%","-36%")
) %>% 
  select(Team = Club, Payroll = `2021`) %>% 
  mutate(
    Payroll = str_sub(Payroll, 2, -1),
    Payroll = str_remove_all(Payroll, ","),
    Payroll = as.double(Payroll)
  )

write_csv(df_payroll, "data/derived/payroll_2021.csv")
